# docs/Enterprise_notebooklm_openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise NotebookLM API
  version: "1.0.0"
  description: >
    API contract cho hệ thống Enterprise NotebookLM (LLM + RAG).
    Bao gồm Workspace Management, Knowledge Base/Document Versioning (Document Understanding qua LlamaParse
    + Hierarchical Chunking), Intelligent Search, AI Chat với Citation (Confidence Engine + Event-driven
    Micro Agents: Rewrite/Graph/SQL), AI Summary, Information Extraction, Multi-document Comparison,
    Report Generation, và Observability.
    Ánh xạ FR1–FR14 (Business_Context.md) và schema v3 (database-design-enterprise-notebooklm.md).
servers:
  - url: https://api.notebooklm.example.com/v1
    description: Production
  - url: https://staging-api.notebooklm.example.com/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: Danh sách Workspace mà user hiện tại có quyền truy cập
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceListResponse" }
    post:
      tags: [Workspaces]
      summary: Tạo Workspace mới (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "201":
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
...
  /workspaces/{workspaceId}/search:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    post:
      tags: [Search]
      summary: Tìm kiếm ngữ nghĩa (Hybrid Retrieval Vector+BM25+KG, ghi search_history)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query_text]
              properties:
                query_text: { type: string }
                filters:
                  type: object
                  description: Lọc theo file_type, khoảng thời gian, tag...
                top_k: { type: integer, default: 10 }
      responses:
        "200":
          description: Kết quả đã rerank
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResultResponse" }
  /workspaces/{workspaceId}/search/history:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    get:
      tags: [Search]
      summary: Lịch sử tìm kiếm của user hiện tại trong Workspace
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SearchHistoryItem" }
  /workspaces/{workspaceId}/search/history/{historyId}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - name: historyId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Search]
      summary: Ghi nhận tài liệu được click từ một lần tìm kiếm (clicked_document_id)
      description: >
        Idempotent — gọi lại với cùng clicked_document_id không gây tác dụng phụ.
        Chỉ chủ sở hữu history (user hiện tại) mới được cập nhật.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clicked_document_id]
              properties:
                clicked_document_id: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchHistoryItem" }
        "404":
          description: History không tồn tại hoặc không thuộc user hiện tại
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ---------------------------------------------------------------------
  # CHAT (FR4, FR5, FR10, FR11)
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
...
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkspaceListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
    SearchResultResponse:
      type: object
      properties:
        history_id:
          type: string
          format: uuid
          description: >
            Id bản ghi search_history vừa tạo — dùng cho PATCH
            /search/history/{historyId} khi user click một kết quả.
        results_count: { type: integer }
        results:
          type: array
          items:
            type: object
            properties:
              chunk_id: { type: string, format: uuid, nullable: true }
              entity_id: { type: string, format: uuid, nullable: true }
              document_id: { type: string, format: uuid }
              document_title: { type: string, nullable: true }
              text_snippet: { type: string }
              retrieval_method:
                type: string
                enum: [vector, bm25, knowledge_graph, rerank]
              score: { type: number, format: float }
              rank: { type: integer }
              page_number: { type: integer, nullable: true }
              location:
                $ref: "#/components/schemas/ContentLocation"
                nullable: true
    SearchHistoryItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        query_text: { type: string }
        filters: { type: object }
        results_count: { type: integer }
        clicked_document_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
...
    ContentLocation:
      type: object
      description: Vị trí nội dung — page_number XOR section_index theo file_type (FR5)
      properties:
        page_number: { type: integer, nullable: true }
        section_index: { type: integer, nullable: true }
        section_title: { type: string, nullable: true }
...